pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-eval-pipeline"
version = "0.1.0"
description = "Ollama-based clinical LLM reproducibility evaluation pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "pandas>=2.2.2",
  "numpy>=1.26.4",
  "requests>=2.32.3",
  "matplotlib>=3.9.0",
  "seaborn>=0.13.2",
  "pyarrow>=16.1.0",
  "tabulate>=0.9.0",
]

[project.scripts]
clinical-eval = "clinical_eval_pipeline.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
